refactor: extract the grpc mod
